+++ controllers/AdminController.php
@@ -4,16 +4,20 @@ namespace humhub\modules\fcmPush\controllers;
 
 use humhub\modules\admin\components\Controller;
 use humhub\modules\fcmPush\models\ConfigureForm;
+use humhub\modules\fcmPush\Module;
 use Yii;
 
+/**
+ *
+ * @property Module $module
+ */
 class AdminController extends Controller
 {
 
     public function actionIndex()
     {
 
-        $model = new ConfigureForm();
-        $model->loadSettings();
+        $model = $this->module->getConfigureForm();
 
         if ($model->load(Yii::$app->request->post()) && $model->validate() && $model->saveSettings()) {
             $this->view->saved();