+++ controllers/TokenController.php
@@ -4,12 +4,11 @@
 namespace humhub\modules\fcmPush\controllers;
 
 use humhub\components\Controller;
-use humhub\modules\fcmPush\models\FcmUser;
+use humhub\modules\fcmPush\services\TokenService;
 use Yii;
 
 class TokenController extends Controller
 {
-
     public function actionUpdate()
     {
         $this->forcePostRequest();
@@ -21,20 +20,8 @@ class TokenController extends Controller
 
         $token = (string)Yii::$app->request->post('token');
 
-        $fcmUser = FcmUser::findOne(['token' => $token]);
-        if ($fcmUser !== null && $fcmUser->user_id !== Yii::$app->user->id) {
-            $fcmUser->delete();
-            $fcmUser = null;
-        }
-
-        if ($fcmUser === null) {
-            $fcmUser = new FcmUser();
-            $fcmUser->user_id = Yii::$app->user->id;
-            $fcmUser->token = $token;
-        }
-
-
-        return $this->asJson(['success' => ($fcmUser->save())]);
+        $tokenService = new TokenService();
+        return $this->asJson(['success' => ($tokenService->storeTokenForUser(Yii::$app->user->getIdentity(), $token))]);
     }
 
 }