+++ docs/CHANGELOG.md
@@ -1,6 +1,15 @@
-1.0.1  (January 21, 2021)
+2.0.0  (TBD)
 -------------------------
 
+:warning: This major release an updated module configuration! 
+
+- Enh: Switched to new Google Firebase API
+- Enh: Added CLI Test Command
+- Enh: Bundle JS File instead of use Google CDN
+
+
+1.0.1  (January 21, 2021)
+-------------------------
 - Fix: Initialize firebase only if module was configured
 - Enh: Updated CHANGELOG